# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SpaceArt API",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://api-spaceart.local/api/"
        }
    ],
    "paths": {
        "/agreement": {
            "get": {
                "tags": [
                    "Agreement"
                ],
                "summary": "List agreements",
                "description": "Fetch user's agreements on database",
                "operationId": "b0b7da895d6337b674f1bd9aa691dcb3",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Limit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agreements found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Agreement"
                                            }
                                        },
                                        "fails": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agreement"
                ],
                "summary": "Store agreement",
                "description": "Store agreement on database",
                "operationId": "f10a4ab3169d1c99602c0cbb796fb1bc",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AgreementStore"
                },
                "responses": {
                    "200": {
                        "description": "Agreements was created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Agreement"
                                            },
                                            "maxItems": 1
                                        },
                                        "fails": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/agreement/{id}": {
            "get": {
                "tags": [
                    "Agreement"
                ],
                "summary": "Fetch agreement",
                "description": "Fetch an unique agreement on database",
                "operationId": "9fa8fb46551571eb6faa2efa87bc758e",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnAgreement"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agreement"
                ],
                "summary": "Update agreement",
                "description": "Update agreement on database",
                "operationId": "02e650c7a48dc10b28c4617ab39b57d1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AgreementUpdate"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReturnAgreement"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agreement"
                ],
                "summary": "Delete agreement",
                "description": "Delete agreement on database",
                "operationId": "6fe0d6aab93ac2f9f1c3cec475827b1d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/agreement/{id}/update": {
            "post": {
                "tags": [
                    "Agreement"
                ],
                "summary": "[PUT]::/agreement/{id} alias",
                "description": "Redirect request to [PUT]::/agreement/{id} alias",
                "operationId": "917f24a664563a08e429246d8456463d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirected to [PUT]::/agreement/{id}"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/agreement/{id}/delete": {
            "post": {
                "tags": [
                    "Agreement"
                ],
                "summary": "[DELETE]::/agreement/{id} alias",
                "description": "Redirect request to [DELETE]::/agreement/{id} alias",
                "operationId": "c204776d0d5f93ef877ff6c54a0d2013",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Id"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirected to [DELETE]::/agreement/{id}"
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
        "/auth": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "8e2b0b8ec2779a0f85d9e8e68d5f00a3",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Auth"
                },
                "responses": {
                    "200": {
                        "description": "Has been authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Authentication failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "default": "User not authenticated."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "": {
                                                        "default": "Email or password is incorrect."
                                                    }
                                                },
                                                "type": "object",
                                                "minItems": 1
                                            }
                                        },
                                        "fails": {
                                            "type": "boolean",
                                            "default": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "default": "User logout."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "$ref": "#/components/securitySchemes/Sanctum"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Agreement": {
                "description": "Schema of Model Agreement",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "enterprise": {
                        "$ref": "#/components/schemas/User"
                    },
                    "artist": {
                        "$ref": "#/components/schemas/User"
                    },
                    "art": {
                        "$ref": "#/components/schemas/Art"
                    },
                    "note": {
                        "description": "Description of service",
                        "type": "string",
                        "example": "Rock musical show for 2 hours"
                    },
                    "date": {
                        "description": "Day of service",
                        "type": "string",
                        "format": "date",
                        "example": "01/01/2025"
                    },
                    "start_time": {
                        "description": "Hour of service",
                        "type": "string",
                        "format": "date",
                        "example": "18:30"
                    },
                    "end_time": {
                        "description": "Hour of finish service",
                        "type": "string",
                        "format": "date",
                        "example": "20:30"
                    },
                    "status": {
                        "type": "enum",
                        "enum": [
                            "send",
                            "accepted",
                            "refused",
                            "canceled"
                        ]
                    }
                },
                "type": "object"
            },
            "Art": {
                "description": "Schema of Model Art",
                "properties": {
                    "name": {
                        "type": "enum",
                        "enum": [
                            "acting",
                            "dance",
                            "music",
                            "painting",
                            "sculpture"
                        ]
                    },
                    "artists_count": {
                        "description": "number of relationed artists",
                        "type": "integer"
                    },
                    "agreements_count": {
                        "description": "number of relationed agreements",
                        "type": "integer"
                    },
                    "selectives_count": {
                        "description": "number of relationed selectives",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "ReturnAgreement": {
                "description": "Response is successful",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Agreement"
                                    },
                                    "maxItems": 1
                                },
                                "fails": {
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "204": {
                "description": "Resource was disabled",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "fails": {
                                    "type": "boolean",
                                    "default": "false"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "parameters": {
            "Limit": {
                "name": "limit",
                "in": "query",
                "description": "Limit per page",
                "style": "form",
                "schema": {
                    "type": "integer",
                    "nullable": true
                }
            },
...
            "Id": {
                "name": "id",
                "in": "path",
                "description": "Resource id",
                "style": "form",
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "requestBodies": {
            "AgreementStore": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "enterprise_id",
                                "artist_id",
                                "price",
                                "note",
                                "date",
                                "start_time",
                                "end_time"
                            ],
                            "properties": {
                                "enterprise_id": {
                                    "type": "integer"
                                },
                                "artist_id": {
                                    "type": "integer"
                                },
                                "price": {
                                    "type": "number",
                                    "format": "^\\d*(\\.\\d*)?$",
                                    "example": "350.99"
                                },
                                "note": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "d/m/Y",
                                    "example": "31/12/2005"
                                },
                                "start_time": {
                                    "type": "string",
                                    "format": "H:i",
                                    "example": "00:00"
                                },
                                "end_time": {
                                    "type": "string",
                                    "format": "H:i",
                                    "example": "22:00"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "AgreementUpdate": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "status"
                            ],
                            "properties": {
                                "note": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string",
                                    "format": "d/m/Y",
                                    "example": "31/12/2005"
                                },
                                "start_time": {
                                    "type": "string",
                                    "format": "H:i",
                                    "example": "00:00"
                                },
                                "end_time": {
                                    "type": "string",
                                    "format": "H:i",
                                    "example": "22:00"
                                },
                                "status": {
                                    "type": "enum",
                                    "enum": [
                                        "send",
                                        "accepted",
                                        "refused",
                                        "canceled"
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Auth": {
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "email",
                                "password",
                                "device_name"
                            ],
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string",
                                    "minLength": 8
                                },
                                "device_name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Sanctum": {
                "type": "apiKey",
                "scheme": {
                    "bearerToken": []
                }
            }
        }
    }
}